[vscode] Display non 200 error nicely (#9785)
