+++ firebase-vscode/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## NEXT
 
-- Render GraphQL error debug details and error code in the execution panel nicely. (#9769)
+- Render GraphQL error debug details and error code in the execution panel nicely. (#9769, #9785)
 - Language server now recognizes multiple schemas. (#9768)
 
 ## 2.0.1
