+++ firebase-vscode/common/messaging/protocol.ts
@@ -10,6 +10,7 @@ import { RCData } from "../../../src/rc";
 import { EmulatorsStatus, RunningEmulatorInfo } from "./types";
 import { ExecutionResult } from "graphql";
 import { SerializedError } from "../error";
+import { GraphqlError, GraphqlResponseError } from "../dataconnect/types";
 
 export enum AuthParamsKind {
   ADMIN = "admin",
@@ -138,9 +139,20 @@ export interface WebviewToExtensionParamsMap {
 export interface DataConnectResults {
   displayName: string;
   query: string;
-  results: ExecutionResult | SerializedError;
   variables: string;
   auth: AuthParams;
+  results: ExecutionResults;
+}
+
+// If non-200 status: respErr and errors from `details` is set
+// If 200 status:
+//   - success: only data is set
+//   - request error: only errors is set
+//   - field error: both data and errors are set
+export interface ExecutionResults {
+  data?: any; // data can be any valid JSON value.
+  gqlErrors?: GraphqlError[];
+  respErr: GraphqlResponseError | SerializedError | undefined;
 }
 
 export type ValueOrError<T> =
